[Bronze IV] Title: A Simple Problem., Time: 136 ms, Memory: 22888 KB -BaekjoonHub

+++ 백준/Bronze/15372. A Simple Problem．/A Simple Problem．.cs
@@ -9,7 +9,7 @@ int t = int.Parse(streamReader.ReadLine());
 for (int i = 0; i < t ; i++)
 {
     int n = int.Parse(streamReader.ReadLine());
-    stringBuilder.AppendLine(((BigInteger)n * n).ToString());
+    stringBuilder.AppendLine(((long)n * n).ToString());
 }
 
 stringWriter.WriteLine(stringBuilder.ToString());
